pyproject.toml
[tool.poetry]
name = "maho"
version = "0.3.0"
description = "Discord Bot with various utility functions"
authors = ["Rain <rainmjp@gmail.com>"]
license = "MIT"
include = ["src/maho/cogs/static/*", "src/maho/Logs/*"]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
pyyaml = "^6.0"
peewee = "^3.15.2"
py-cord = {version="^2.0.0", extras=["speed"]}

[tool.poetry.dev-dependencies]
black = {version = "^22.8.0", allow-prereleases = true}
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
pytest-env = "^0.6.2"
pytest-asyncio = "^0.19.0"

[tool.poetry.scripts]
config = "scripts:config"
start = "scripts:start"
test = "scripts:test"
init-db = "scripts:init_db"

[tool.coverage.run]
source = ["src/maho"]
omit = ["src/maho/main.py"]

[tool.coverage.report]
fail_under = 80

[tool.isort]
balanced_wrapping = true
combine_as_imports = true
default_section = "THIRDPARTY"
known_maho = "maho"
length_sort = false
multi_line_output = 6
sections = "FUTURE,STDLIB,THIRDPARTY,MAHO,LOCALFOLDER"
skip = ".git,.venv,.eggs,.mypy_cache,__pycache__,docs,build,dist,data"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src", "tests"]
addopts = [
    "--cov",
    "--cov-report=term",
]
env = ["BOT_DB=:memory:"]
log_level = "ERROR"

[build-system]
requires = ["poetry>=1.1"]
build-backend = "poetry.masonry.api"
